Bugfix (workaround) for regex search matching deleted text
Current implementation of Onigmo on Scintilla uses Scintilla's RangePointer() to get text range references. However, RangePointer() was designed with text insertion in mind, and as such might return pointers passed the real end of the text, into a "gap" that Scintilla maintains for inexpensive text appends and that may contain invalid data, including deleted text. Under certain conditions, a regex search might turn up an invalid match when such data is included as part of the subject of the search, while being non-existent in the document itself.

It's not a Scintilla's bug per-se, but rather a misunderstanding of the way Scintilla works, and a broken contract between the API and the caller.